fix CI problem and reviews
