+++ cloud/pkg/apis/rules/v1/register.go
@@ -14,8 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// NOTE: Boilerplate only.  Ignore this file.
-
 // Package v1 contains API Schema definitions for the rules v1 API group
 // +k8s:openapi-gen=true
 // +k8s:deepcopy-gen=package,register
