+++ cloud/pkg/router/utils/path.go
@@ -8,62 +8,49 @@ import (
 )
 
 const (
-	ParamRegex = "{[-A-Za-z0-9+&@#%?=~_|!:,.;]+}"
-	pathRegex  = "[-A-Za-z0-9+&@#%?=~_|!:,.;]+"
+	pathRegex = "[-A-Za-z0-9+&@#%?=~_|!:,.;]+"
+	tailRegex = "/?"
 )
 
-// URLToURLRegex return url regex and replace {} in url
+var paramRegex = regexp.MustCompile("{" + pathRegex + "}")
+
+// URLToURLRegex return url regex and replace {} in url,e.g.,/abc/{Aa1} to /abc/[-A-Za-z0-9+&@#%?=~_|!:,.;]+/?
 func URLToURLRegex(url string) string {
-	paramRegex := regexp.MustCompile(ParamRegex)
 	params := paramRegex.FindAllString(url, -1)
 	for _, param := range params {
 		url = strings.Replace(url, param, pathRegex, -1)
 	}
-	url = url + "/{0,1}"
+	url = url + tailRegex
 	return url
 }
 
-// isMatch return true if the request uri math rule using regex
+// IsMatch return true if the path match rule using regex
 func IsMatch(reg, path string) bool {
 	match, err := regexp.MatchString(URLToURLRegex(reg), path)
 	if err != nil {
 		klog.Errorf("failed to validate res %s and reqPath %s, err: %v", reg, path, err)
 		return false
 	}
-	if match {
-		return true
-	}
-	return false
+	return match
 }
 
-// RuleContains return true if rule 1 contains rule 2
-func RuleContains(rulePath []string, rule2Path []string) bool {
-	if len(rulePath) == 0 {
+// RuleContains return true if rule 1 contains rule 2, e.g., path /a contains /a/b
+func RuleContains(rulePath, rule2Path string) bool {
+	path1 := strings.Split(rulePath, "/")
+	path2 := strings.Split(rule2Path, "/")
+	if len(path1) == 0 {
 		return true
 	}
 
-	if len(rule2Path) == 0 {
+	if len(path2) == 0 {
 		return false
 	}
 
-	// rule1[i] contains rule2[i] when rule1[i] = {} or rule1[i] == rule2[i]
-	match1, err := regexp.MatchString(URLToURLRegex(rulePath[0]), rule2Path[0])
-	if err != nil {
-		klog.Errorf("failed to validate rulePath %s and rule2Path %s, err: %v", rulePath[0], rule2Path[0], err)
-		return false
+	for i := 0; i < len(path1) && i < len(path2); i++ {
+		// rule1[i] contains rule2[i] when rule1[i] = {} or rule1[i] == rule2[i]
+		if path1[i] != path2[i] && URLToURLRegex(path1[i]) != URLToURLRegex(path2[i]) {
+			return false
+		}
 	}
-	match2, err := regexp.MatchString(ParamRegex, rule2Path[0])
-	if err != nil {
-		klog.Errorf("failed to validate rule2Path %s, err: %v", rule2Path[0], err)
-		return false
-	}
-	match3, err := regexp.MatchString(ParamRegex, rulePath[0])
-	if err != nil {
-		klog.Errorf("failed to validate rulePath %s, err: %v", rulePath[0], err)
-		return false
-	}
-	if rulePath[0] == rule2Path[0] || match1 || (match2 && match3) {
-		return RuleContains(rulePath[1:], rule2Path[1:])
-	}
-	return false
+	return true
 }
