+++ staging/src/github.com/kubeedge/beehive/pkg/core/model/message.go
@@ -23,7 +23,7 @@ const (
 	ResourceTypePodStatus    = "podstatus"
 	ResourceTypeNodeStatus   = "nodestatus"
 	ResourceTypeRule         = "rule"
-	ResourceTypeRuleEndpoint = "rule-endpoint"
+	ResourceTypeRuleEndpoint = "ruleendpoint"
 )
 
 // Message struct
