+++ tests/e2e/utils/common.go
@@ -41,7 +41,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/kubeedge/kubeedge/cloud/pkg/apis/devices/v1alpha2"
-	rulesv1 "github.com/kubeedge/kubeedge/cloud/pkg/apis/rules/v1"
 	"github.com/kubeedge/kubeedge/common/constants"
 	"github.com/kubeedge/viaduct/pkg/api"
 )
@@ -896,157 +895,6 @@ func CompareTwin(deviceTwin map[string]*MsgTwin, expectedDeviceTwin map[string]*
 	return true
 }
 
-// GetRuleList to get the rule list and verify whether the contents of the rule matches with what is expected
-func GetRuleList(list *rulesv1.RuleList, getRuleAPI string, expectedRule *rulesv1.Rule) ([]rulesv1.Rule, error) {
-	resp, err := SendHTTPRequest(http.MethodGet, getRuleAPI)
-	defer resp.Body.Close()
-	contents, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		Fatalf("HTTP Response reading has failed: %v", err)
-		return nil, err
-	}
-	err = json.Unmarshal(contents, &list)
-	if err != nil {
-		Fatalf("Unmarshal HTTP Response has failed: %v", err)
-		return nil, err
-	}
-	if expectedRule != nil {
-		modelExists := false
-		for _, rule := range list.Items {
-			if expectedRule.ObjectMeta.Name == rule.ObjectMeta.Name {
-				modelExists = true
-				if !reflect.DeepEqual(expectedRule.TypeMeta, rule.TypeMeta) ||
-					expectedRule.ObjectMeta.Namespace != rule.ObjectMeta.Namespace ||
-					!reflect.DeepEqual(expectedRule.Spec, rule.Spec) {
-					return nil, errors.New("The rule is not matching with what was expected")
-				}
-			}
-		}
-		if !modelExists {
-			return nil, errors.New("The requested rule is not found")
-		}
-	}
-	return list.Items, nil
-}
-
-// HandleRule to handle rule.
-func HandleRule(operation, apiserver, UID, sourceType, targetType string) (bool, int) {
-	var req *http.Request
-	var err error
-	var body io.Reader
-
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-	}
-	client := &http.Client{
-		Transport: tr,
-	}
-
-	switch operation {
-	case http.MethodPost:
-		body := NewRule(sourceType, targetType)
-		respBytes, err := json.Marshal(body)
-		if err != nil {
-			Fatalf("Marshalling body failed: %v", err)
-		}
-		req, err = http.NewRequest(http.MethodPost, apiserver, bytes.NewBuffer(respBytes))
-		req.Header.Set("Content-Type", "application/json")
-	case http.MethodDelete:
-		req, err = http.NewRequest(http.MethodDelete, apiserver+UID, body)
-		req.Header.Set("Content-Type", "application/json")
-	}
-	if err != nil {
-		// handle error
-		Fatalf("Frame HTTP request failed: %v", err)
-		return false, 0
-	}
-	t := time.Now()
-	resp, err := client.Do(req)
-	if err != nil {
-		// handle error
-		Fatalf("HTTP request is failed :%v", err)
-		return false, 0
-	}
-	contents, err := ioutil.ReadAll(resp.Body)
-	Infof("%s %s %v  %v in %v", req.Method, req.URL, resp.Status, string(contents), time.Since(t))
-	return true, resp.StatusCode
-}
-
-// HandleRuleEndpoint to handle ruleendpoint.
-func HandleRuleEndpoint(operation string, apiserver string, UID string, endpointType string) (bool, int) {
-	var req *http.Request
-	var err error
-	var body io.Reader
-
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-	}
-	client := &http.Client{
-		Transport: tr,
-	}
-
-	switch operation {
-	case http.MethodPost:
-		body := NewRuleEndpoint(endpointType)
-		respBytes, err := json.Marshal(body)
-		if err != nil {
-			Fatalf("Marshalling body failed: %v", err)
-		}
-		req, err = http.NewRequest(http.MethodPost, apiserver, bytes.NewBuffer(respBytes))
-		req.Header.Set("Content-Type", "application/json")
-	case http.MethodDelete:
-		req, err = http.NewRequest(http.MethodDelete, apiserver+UID, body)
-		req.Header.Set("Content-Type", "application/json")
-	}
-	if err != nil {
-		// handle error
-		Fatalf("Frame HTTP request failed: %v", err)
-		return false, 0
-	}
-	t := time.Now()
-	resp, err := client.Do(req)
-	if err != nil {
-		// handle error
-		Fatalf("HTTP request is failed :%v", err)
-		return false, 0
-	}
-	Infof("%s %s %v in %v", req.Method, req.URL, resp.Status, time.Since(t))
-	return true, resp.StatusCode
-}
-
-// GetRuleEndpointList to get the rule endpoint list and verify whether the contents of the ruleendpoint matches with what is expected
-func GetRuleEndpointList(list *rulesv1.RuleEndpointList, getRuleEndpointAPI string, expectedRule *rulesv1.RuleEndpoint) ([]rulesv1.RuleEndpoint, error) {
-	resp, err := SendHTTPRequest(http.MethodGet, getRuleEndpointAPI)
-	defer resp.Body.Close()
-	contents, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		Fatalf("HTTP Response reading has failed: %v", err)
-		return nil, err
-	}
-	err = json.Unmarshal(contents, &list)
-	if err != nil {
-		Fatalf("Unmarshal HTTP Response has failed: %v", err)
-		return nil, err
-	}
-	if expectedRule != nil {
-		exists := false
-		for _, ruleEndpoint := range list.Items {
-			if expectedRule.ObjectMeta.Name == ruleEndpoint.ObjectMeta.Name {
-				exists = true
-				if !reflect.DeepEqual(expectedRule.TypeMeta, ruleEndpoint.TypeMeta) ||
-					expectedRule.ObjectMeta.Namespace != ruleEndpoint.ObjectMeta.Namespace ||
-					!reflect.DeepEqual(expectedRule.Spec, ruleEndpoint.Spec) {
-					return nil, errors.New("The ruleendpoint is not matching with what was expected")
-				}
-			}
-		}
-		if !exists {
-			return nil, errors.New("The requested ruleendpoint is not found")
-		}
-	}
-	return list.Items, nil
-}
-
 func SendMsg(url string, message []byte) (bool, int) {
 	var req *http.Request
 	var err error
@@ -1060,14 +908,14 @@ func SendMsg(url string, message []byte) (bool, int) {
 	req, err = http.NewRequest(http.MethodPost, url, bytes.NewBuffer(message))
 	if err != nil {
 		// handle error
-		Fatalf("Frame HTTP request failed: %v", err)
+		Fatalf("Frame HTTP request failed, request: %s, reason: %v", req.URL.String(), err)
 		return false, 0
 	}
 	t := time.Now()
 	resp, err := client.Do(req)
 	if err != nil {
 		// handle error
-		Fatalf("HTTP request is failed :%v", err)
+		Fatalf("HTTP request is failed: %v", err)
 		return false, 0
 	}
 	Infof("%s %s %v in %v", req.Method, req.URL, resp.Status, time.Since(t))
